Changed Organization & deleted unnecessary files
