+++ AirRes/src/main/Passenger.java
@@ -13,22 +13,5 @@ public class Passenger {
 	public String getName() {
 		return name;
 	}
-	
-	 @Override
-	    public boolean equals(Object o){
-	           
-	           if(o==null)
-	                  return false;
-	           if(this.getClass()!=o.getClass())
-	                  return false;
-	    
-	           Passenger p=(Passenger)o;
-	           return p.name.equals(this.name);          
-	    }
-	           
-	    @Override
-	    public int hashCode(){
-	           return  name.hashCode();      
-	    }
 
 }
