--- AirRes/src/main/PlaneModel.java
@@ -1,20 +1,0 @@
-package main;
-
-public class PlaneModel {
-
-	private final int seats;
-	private final String ID;
-	
-	public PlaneModel(final int seats, final String ID) {
-		this.seats = seats;
-		this.ID = ID;
-	}
-	
-	
-	public boolean hasSeats() {
-		
-		return false;
-		
-	}
-	
-}
