--- AirRes/src/main/SystemIn.java
@@ -1,7 +1,0 @@
-package main;
-
-public class SystemIn {
-
-	
-	
-}
