+++ AirRes/src/test/HashMapTest.java
@@ -1,0 +1,14 @@
+package test;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class HashMapTest {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
